requirements.txt
tensorflow
keras
numpy
streamlit
pillow
